profile_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
google-cloud-pubsub
